Write a script to import new PNC data from local files and remove references to ACPO data since it's duplicated anyway
