+++ src/download-offence-code-data.ts
@@ -1,11 +1,9 @@
 import cjsDownload from "./cjs-download"
-// import pncDownload from "./pnc-download"
 import pnldDownload from "./pnld-download"
 
 const main = async () => {
   await pnldDownload()
   await cjsDownload()
-  // await pncDownload()
 }
 
 main().then(() => console.log("All data downloaded"))
